Moved and made new requirements.txt with pipreqs
