+++ requirements.txt
@@ -1,211 +1,5 @@
-# This file may be used to create an environment using:
-# $ conda create --name <env> --file <this file>
-# platform: linux-64
-_libgcc_mutex=0.1=conda_forge
-_openmp_mutex=4.5=2_gnu
-alsa-lib=1.2.8=h166bdaf_0
-argcomplete=3.1.1=pyhd8ed1ab_0
-argh=0.27.2=pyhd8ed1ab_0
-asttokens=2.2.1=pyhd8ed1ab_0
-attr=2.5.1=h166bdaf_1
-autopep8=2.0.4=pyhd8ed1ab_0
-backcall=0.2.0=pyh9f0ad1d_0
-backports=1.0=pyhd8ed1ab_3
-backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
-bcbio-gff=0.7.0=pypi_0
-biopython=1.81=py311h2582759_0
-brotli=1.0.9=h166bdaf_9
-brotli-bin=1.0.9=h166bdaf_9
-brotli-python=1.0.9=py311ha362b79_9
-bzip2=1.0.8=h7f98852_4
-ca-certificates=2023.7.22=hbcca054_0
-cairo=1.16.0=hbbf8b49_1016
-certifi=2023.7.22=pyhd8ed1ab_0
-charset-normalizer=3.2.0=pyhd8ed1ab_0
-comm=0.1.4=pyhd8ed1ab_0
-contourpy=1.1.0=py311h9547e67_0
-cycler=0.11.0=pyhd8ed1ab_0
-cython=3.0.0=pypi_0
-dbus=1.13.6=h5008d03_3
-debugpy=1.6.8=py311hb755f60_0
-decorator=5.1.1=pyhd8ed1ab_0
-executing=1.2.0=pyhd8ed1ab_0
-expat=2.5.0=hcb278e6_1
-font-ttf-dejavu-sans-mono=2.37=hab24e00_0
-font-ttf-inconsolata=3.000=h77eed37_0
-font-ttf-source-code-pro=2.038=h77eed37_0
-font-ttf-ubuntu=0.83=hab24e00_0
-fontconfig=2.14.2=h14ed4e7_0
-fonts-conda-ecosystem=1=0
-fonts-conda-forge=1=0
-fonttools=4.42.0=py311h459d7ec_0
-freetype=2.12.1=hca18f0e_1
-gettext=0.21.1=h27087fc_0
-gffpandas=1.2.0=pypi_0
-gffutils=0.12=pyh7cba7a3_0
-glib=2.76.4=hfc55251_0
-glib-tools=2.76.4=hfc55251_0
-graphite2=1.3.13=h58526e2_1001
-gst-plugins-base=1.22.3=he621ea3_0
-gstreamer=1.22.3=h5eee18b_0
-harfbuzz=7.3.0=hdb3a94d_0
-icu=72.1=hcb278e6_0
-idna=3.4=pyhd8ed1ab_0
-igv-reports=1.8.1=pypi_0
-importlib-metadata=6.8.0=pyha770c72_0
-importlib_metadata=6.8.0=hd8ed1ab_0
-intervaltree=3.1.0=pypi_0
-ipykernel=6.25.1=pyh71e2992_0
-ipython=8.14.0=pyh41d4057_0
-jedi=0.19.0=pyhd8ed1ab_0
-jupyter_client=8.3.0=pyhd8ed1ab_0
-jupyter_core=5.3.1=py311h38be061_0
-keyutils=1.6.1=h166bdaf_0
-kiwisolver=1.4.4=py311h4dd048b_1
-krb5=1.20.1=h81ceb04_0
-lame=3.100=h166bdaf_1003
-lcms2=2.15=haa2dc70_1
-ld_impl_linux-64=2.40=h41732ed_0
-lerc=4.0.0=h27087fc_0
-libblas=3.9.0=17_linux64_openblas
-libbrotlicommon=1.0.9=h166bdaf_9
-libbrotlidec=1.0.9=h166bdaf_9
-libbrotlienc=1.0.9=h166bdaf_9
-libcap=2.69=h0f662aa_0
-libcblas=3.9.0=17_linux64_openblas
-libclang=16.0.6=default_h1cdf331_1
-libclang13=16.0.6=default_h4d60ac6_1
-libcups=2.3.3=h36d4200_3
-libdeflate=1.18=h0b41bf4_0
-libedit=3.1.20191231=he28a2e2_2
-libevent=2.1.12=hf998b51_1
-libexpat=2.5.0=hcb278e6_1
-libffi=3.4.2=h7f98852_5
-libflac=1.4.3=h59595ed_0
-libgcc-ng=13.1.0=he5830b7_0
-libgcrypt=1.10.1=h166bdaf_0
-libgfortran-ng=13.1.0=h69a702a_0
-libgfortran5=13.1.0=h15d22d2_0
-libglib=2.76.4=hebfc3b9_0
-libgomp=13.1.0=he5830b7_0
-libgpg-error=1.47=h71f35ed_0
-libiconv=1.17=h166bdaf_0
-libjpeg-turbo=2.1.5.1=h0b41bf4_0
-liblapack=3.9.0=17_linux64_openblas
-libllvm16=16.0.6=h5cf9203_1
-libnsl=2.0.0=h7f98852_0
-libogg=1.3.5=h27cfd23_1
-libopenblas=0.3.23=pthreads_h80387f5_0
-libopus=1.3.1=h7f98852_1
-libpng=1.6.39=h753d276_0
-libpq=15.3=hbcd7760_1
-libsndfile=1.2.0=hb75c966_0
-libsodium=1.0.18=h36c2ea0_1
-libsqlite=3.42.0=h2797004_0
-libstdcxx-ng=13.1.0=hfd8a6a1_0
-libsystemd0=254=h3516f8a_0
-libtiff=4.5.1=h8b53f26_0
-libuuid=2.38.1=h0b41bf4_0
-libvorbis=1.3.7=h9c3ff4c_0
-libwebp-base=1.3.1=hd590300_0
-libxcb=1.15=h0b41bf4_0
-libxkbcommon=1.5.0=h5d7e998_3
-libxml2=2.11.4=h0d562d8_0
-libzlib=1.2.13=hd590300_5
-lz4-c=1.9.4=hcb278e6_0
-matplotlib=3.7.2=py311h38be061_0
-matplotlib-base=3.7.2=py311h54ef318_0
-matplotlib-inline=0.1.6=pyhd8ed1ab_0
-mpg123=1.31.3=hcb278e6_0
-munkres=1.0.7=py_1
-mysql-common=8.0.33=hf1915f5_2
-mysql-libs=8.0.33=hca2cd23_2
-ncurses=6.4=hcb278e6_0
-nest-asyncio=1.5.6=pyhd8ed1ab_0
-nspr=4.35=h27087fc_0
-nss=3.89=he45b914_0
-numpy=1.25.2=py311h64a7726_0
-openjpeg=2.5.0=hfec8fc6_2
-openssl=3.1.2=hd590300_0
-packaging=23.1=pyhd8ed1ab_0
-pandas=2.0.3=py311h320fe9a_1
-parso=0.8.3=pyhd8ed1ab_0
-patsy=0.5.3=pyhd8ed1ab_0
-pcre2=10.40=hc3806b6_0
-pexpect=4.8.0=pyh1a96a4e_2
-pickleshare=0.7.5=py_1003
-pillow=10.0.0=py311h0b84326_0
-pip=23.2.1=pyhd8ed1ab_0
-pixman=0.40.0=h36c2ea0_0
-platformdirs=3.10.0=pyhd8ed1ab_0
-ply=3.11=py_1
-pooch=1.7.0=pyha770c72_3
-prompt-toolkit=3.0.39=pyha770c72_0
-prompt_toolkit=3.0.39=hd8ed1ab_0
-psutil=5.9.5=py311h2582759_0
-pthread-stubs=0.4=h36c2ea0_1001
-ptyprocess=0.7.0=pyhd3deb0d_0
-pulseaudio-client=16.1=hb77b528_4
-pure_eval=0.2.2=pyhd8ed1ab_0
-pycodestyle=2.11.0=pyhd8ed1ab_0
-pyfaidx=0.7.2.1=pyh7cba7a3_1
-pygments=2.16.1=pyhd8ed1ab_0
-pyparsing=3.0.9=pyhd8ed1ab_0
-pyqt=5.15.9=py311hf0fb5b6_4
-pyqt5-sip=12.12.2=py311hb755f60_4
-pysam=0.21.0=pypi_0
-pysocks=1.7.1=pyha2e5f31_6
-python=3.11.4=hab00c5b_0_cpython
-python-dateutil=2.8.2=pyhd8ed1ab_0
-python-tzdata=2023.3=pyhd8ed1ab_0
-python_abi=3.11=3_cp311
-pytz=2023.3=pyhd8ed1ab_0
-pyvcf3=1.0.3=pyhdfd78af_0
-pyzmq=25.1.0=py311h75c88c4_0
-qt-main=5.15.8=h01ceb2d_12
-readline=8.2=h8228510_1
-requests=2.31.0=pyhd8ed1ab_0
-scipy=1.11.1=py311h64a7726_0
-seaborn=0.12.2=hd8ed1ab_0
-seaborn-base=0.12.2=pyhd8ed1ab_0
-setuptools=68.0.0=pyhd8ed1ab_0
-simplejson=3.19.1=py311h2582759_0
-sip=6.7.11=py311hb755f60_0
-six=1.16.0=pyh6c4a22f_0
-sortedcontainers=2.4.0=pypi_0
-stack_data=0.6.2=pyhd8ed1ab_0
-statsmodels=0.14.0=py311h1f0f07a_1
-tk=8.6.12=h27826a3_0
-toml=0.10.2=pyhd8ed1ab_0
-tomli=2.0.1=pyhd8ed1ab_0
-tornado=6.3.2=py311h459d7ec_0
-traitlets=5.9.0=pyhd8ed1ab_0
-typing-extensions=4.7.1=hd8ed1ab_0
-typing_extensions=4.7.1=pyha770c72_0
-tzdata=2023c=h71feb2d_0
-urllib3=2.0.4=pyhd8ed1ab_0
-wcwidth=0.2.6=pyhd8ed1ab_0
-wheel=0.41.1=pyhd8ed1ab_0
-xcb-util=0.4.0=hd590300_1
-xcb-util-image=0.4.0=h8ee46fc_1
-xcb-util-keysyms=0.4.0=h8ee46fc_1
-xcb-util-renderutil=0.3.9=hd590300_1
-xcb-util-wm=0.4.1=h8ee46fc_1
-xkeyboard-config=2.39=hd590300_0
-xorg-kbproto=1.0.7=h7f98852_1002
-xorg-libice=1.1.1=hd590300_0
-xorg-libsm=1.2.4=h7391055_0
-xorg-libx11=1.8.6=h8ee46fc_0
-xorg-libxau=1.0.11=hd590300_0
-xorg-libxdmcp=1.1.3=h7f98852_0
-xorg-libxext=1.3.4=h0b41bf4_2
-xorg-libxrender=0.9.11=hd590300_0
-xorg-renderproto=0.11.1=h7f98852_1002
-xorg-xextproto=7.3.0=h0b41bf4_1003
-xorg-xf86vidmodeproto=2.3.1=h7f98852_1002
-xorg-xproto=7.0.31=h7f98852_1007
-xz=5.2.6=h166bdaf_0
-zeromq=4.3.4=h9c3ff4c_1
-zipp=3.16.2=pyhd8ed1ab_0
-zlib=1.2.13=hd590300_5
-zstd=1.5.2=hfc55251_7
+argcomplete==3.1.1
+numpy==1.25.2
+pandas==2.0.3
+pysam==0.21.0
+pytest==7.4.2
